MacOS Intel machine deployments are updated

+++ deployability/modules/allocation/vagrant/instance.py
@@ -105,13 +105,10 @@ class VagrantInstance(Instance):
         if str(self.host_identifier) == "macstadium":
             logger.debug(f"Deleting remote directory {self.host_instance_dir}")
             VagrantUtils.remote_command(f"sudo rm -rf {self.host_instance_dir}", self.remote_host_parameters)
-            logger.debug(f"Killing remote process on port {self.ssh_port}")
-            proccess = VagrantUtils.remote_command(f"sudo lsof -Pi :{self.ssh_port} -sTCP:LISTEN -t", self.remote_host_parameters)
-            VagrantUtils.remote_command(f"sudo kill -9 {proccess}", self.remote_host_parameters)
-        if str(self.host_identifier) == "black_mini":
-            logger.debug(f"Deleting remote directory {self.host_instance_dir}")
-            VagrantUtils.remote_command(f"sudo rm -rf {self.host_instance_dir}", self.remote_host_parameters)
-
+            if self.arch == 'arm64':
+                logger.debug(f"Killing remote process on port {self.ssh_port}")
+                proccess = VagrantUtils.remote_command(f"sudo lsof -Pi :{self.ssh_port} -sTCP:LISTEN -t", self.remote_host_parameters)
+                VagrantUtils.remote_command(f"sudo kill -9 {proccess}", self.remote_host_parameters)
 
     def status(self) -> str:
         """
@@ -209,7 +206,7 @@ class VagrantInstance(Instance):
         if isinstance(command, str):
             command = [command]
         if self.platform == 'macos':
-            cmd = f"sudo VAGRANT_CWD={self.host_instance_dir} /usr/local/bin/vagrant " + ' '.join(command)
+            cmd = f"sudo {self.host_instance_dir}/vagrant_script.sh " + ' '.join(command)
             output = VagrantUtils.remote_command(cmd, self.remote_host_parameters)
             return output
         else:
